% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.create_progress_bar}
\alias{.create_progress_bar}
\title{Create progress bar if requested and package available}
\usage{
.create_progress_bar(show = TRUE, total = 10, format = "[:bar] :percent :etas")
}
\arguments{
\item{show}{Logical. Should progress bar be created?}

\item{total}{Total number of steps}

\item{format}{Progress bar format string}
}
\value{
Progress bar object or NULL
}
\description{
Create progress bar if requested and package available
}
\keyword{internal}
