% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.get_pkg_env}
\alias{.get_pkg_env}
\title{Get package environment for testing/debugging}
\usage{
.get_pkg_env()
}
\value{
Package environment
}
\description{
Get package environment for testing/debugging
}
\keyword{internal}
