% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.is_initialized}
\alias{.is_initialized}
\title{Check if package is properly initialized}
\usage{
.is_initialized()
}
\value{
Logical indicating initialization status
}
\description{
Check if package is properly initialized
}
\keyword{internal}
