% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attachment message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name (unused)}

\item{pkgname}{Package name (unused)}
}
\description{
Package attachment message
}
\keyword{internal}
