% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize package environment and check dependencies}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name (unused)}

\item{pkgname}{Package name (unused)}
}
\description{
Initialize package environment and check dependencies
}
\keyword{internal}
