% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.tick_progress}
\alias{.tick_progress}
\title{Tick progress bar safely}
\usage{
.tick_progress(pb)
}
\arguments{
\item{pb}{Progress bar object (can be NULL)}
}
\description{
Tick progress bar safely
}
\keyword{internal}
