% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_calculate_optimal_crs}
\alias{euiss_calculate_optimal_crs}
\title{Calculate optimal CRS for the area of interest}
\usage{
euiss_calculate_optimal_crs(countries_crop)
}
\arguments{
\item{countries_crop}{Country boundaries}
}
\value{
CRS string
}
\description{
Calculate optimal CRS for the area of interest
}
