% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{euiss_cartogram_compare}
\alias{euiss_cartogram_compare}
\title{Create multiple cartogram variants for comparison}
\usage{
euiss_cartogram_compare(
  sf_data,
  weight,
  types = c("cont", "dorling"),
  k_values = c(0.05, 0.1, 0.2),
  save_results = FALSE,
  output_dir = "cartograms",
  verbose = TRUE
)
}
\arguments{
\item{sf_data}{Simple Features (sf) object.}

\item{weight}{Weight variable name (unquoted).}

\item{types}{Vector of cartogram types to create (default: c("cont", "dorling")).}

\item{k_values}{Vector of k values to test (default: c(0.05, 0.1, 0.2)).}

\item{save_results}{Save results to files? (default: FALSE)}

\item{output_dir}{Directory for saved results (default: "cartograms").}

\item{verbose}{Print progress? (default: TRUE)}
}
\value{
Named list of cartogram sf objects.
}
\description{
Creates several cartogram variants using different methods and parameters.
Useful for exploring which approach works best for your data.
}
\examples{
\dontrun{
countries <- euiss_gisco(res = "20") \%>\% slice_head(n = 5)
countries$pop <- runif(nrow(countries), 1e6, 1e8)

cartograms <- euiss_cartogram_compare(
  countries,
  weight = pop,
  types = c("cont", "dorling")
)
}

}
