% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-join.R
\name{euiss_code_match}
\alias{euiss_code_match}
\title{Country code matching helper}
\usage{
euiss_code_match(df, sf, location_col, sf_code_col, verbose)
}
\description{
Country code matching helper
}
\keyword{internal}
