% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_configure_basemap_paths}
\alias{euiss_configure_basemap_paths}
\title{Configure basemap data paths with intelligent fallbacks}
\usage{
euiss_configure_basemap_paths(data_path = NULL)
}
\arguments{
\item{data_path}{Optional user-specified data path}
}
\value{
List of configured paths for different data sources
}
\description{
Configure basemap data paths with intelligent fallbacks
}
