% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_create_basemap_plot}
\alias{euiss_create_basemap_plot}
\title{Create the final basemap plot}
\usage{
euiss_create_basemap_plot(
  countries,
  raster,
  roads,
  urban,
  urban_labels,
  lakes,
  target_country,
  ras_name,
  crs,
  colors
)
}
\arguments{
\item{countries}{Country boundaries}

\item{raster}{Raster elevation data}

\item{roads}{Roads data}

\item{urban}{Urban areas}

\item{urban_labels}{Urban labels}

\item{lakes}{Lakes data}

\item{target_country}{Target country name}

\item{ras_name}{Raster variable name}

\item{crs}{Coordinate system}

\item{colors}{Color scheme}
}
\value{
ggplot2 object
}
\description{
Create the final basemap plot
}
