% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-data.R
\name{euiss_download_gisco}
\alias{euiss_download_gisco}
\title{Download GISCO data from Eurostat (REMOVED - Not Implemented)}
\usage{
euiss_download_gisco(...)
}
\arguments{
\item{...}{All parameters (ignored)}
}
\value{
Error with download instructions
}
\description{
This function has been removed as automated downloading was not implemented.
}
\details{
To download GISCO data:
\enumerate{
  \item Visit: \url{https://gisco-services.ec.europa.eu/distribution/v2/countries/}
  \item Download desired resolution files (e.g., ref-countries-2024-20m.shp.zip)
  \item Save to your GISCO data directory
}

Check your data directory with: \code{euiss_get_config()$paths$gisco}
}
