% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-join.R
\name{euiss_exact_match}
\alias{euiss_exact_match}
\title{Exact name matching helper}
\usage{
euiss_exact_match(df, sf, location_col, sf_name_col)
}
\description{
Exact name matching helper
}
\keyword{internal}
