% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-join.R
\name{euiss_fuzzy_match}
\alias{euiss_fuzzy_match}
\title{Fuzzy string matching helper}
\usage{
euiss_fuzzy_match(df, sf, location_col, sf_name_col, threshold)
}
\description{
Fuzzy string matching helper
}
\keyword{internal}
