% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_get_basemap_colors}
\alias{euiss_get_basemap_colors}
\title{Get default EUISS basemap colors}
\usage{
euiss_get_basemap_colors()
}
\value{
Named list of colors for basemap elements
}
\description{
Get default EUISS basemap colors
}
