% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_countries}
\alias{euiss_load_countries}
\title{Load country boundaries data}
\usage{
euiss_load_countries(country, db, buffer, paths)
}
\arguments{
\item{country}{Target country name}

\item{db}{Database source ("gisco" or "ne")}

\item{buffer}{Buffer distance in meters}

\item{paths}{List of data paths}
}
\value{
sf object with cropped country boundaries
}
\description{
Load country boundaries data
}
