% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_lakes}
\alias{euiss_load_lakes}
\title{Load lakes data}
\usage{
euiss_load_lakes(countries_crop, paths)
}
\arguments{
\item{countries_crop}{Cropped boundaries}

\item{paths}{Data paths}
}
\value{
sf object with lakes
}
\description{
Load lakes data
}
