% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_raster_data}
\alias{euiss_load_raster_data}
\title{Load and crop raster elevation data}
\usage{
euiss_load_raster_data(ras, countries_crop, crs, paths)
}
\arguments{
\item{ras}{Raster type}

\item{countries_crop}{Cropped country boundaries}

\item{crs}{Target CRS}

\item{paths}{Data paths}
}
\value{
Tibble with raster data
}
\description{
Load and crop raster elevation data
}
