% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_roads}
\alias{euiss_load_roads}
\title{Load roads data}
\usage{
euiss_load_roads(countries_crop, paths)
}
\arguments{
\item{countries_crop}{Cropped boundaries}

\item{paths}{Data paths}
}
\value{
sf object with roads
}
\description{
Load roads data
}
