% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_urban_areas}
\alias{euiss_load_urban_areas}
\title{Load urban areas data}
\usage{
euiss_load_urban_areas(countries_crop, paths)
}
\arguments{
\item{countries_crop}{Cropped boundaries}

\item{paths}{Data paths}
}
\value{
sf object with urban areas
}
\description{
Load urban areas data
}
