% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_load_urban_labels}
\alias{euiss_load_urban_labels}
\title{Load urban labels/populated places}
\usage{
euiss_load_urban_labels(countries_crop, paths)
}
\arguments{
\item{countries_crop}{Cropped boundaries}

\item{paths}{Data paths}
}
\value{
sf object with place labels
}
\description{
Load urban labels/populated places
}
