% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{euiss_set_data_path}
\alias{euiss_set_data_path}
\title{Set environment variable for data paths}
\usage{
euiss_set_data_path(data_path, permanent = FALSE)
}
\arguments{
\item{data_path}{Base data path}

\item{permanent}{Save to .Renviron file? (default: FALSE)}
}
\description{
Helper function to set environment variables for persistent path configuration.
}
