% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-join.R
\name{euiss_spatial_match}
\alias{euiss_spatial_match}
\title{Spatial intersection matching helper}
\usage{
euiss_spatial_match(df, sf, coords_cols)
}
\description{
Spatial intersection matching helper
}
\keyword{internal}
