% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{euiss_cartogram_compare}
\alias{euiss_cartogram_compare}
\title{Create multiple cartogram variants for comparison}
\usage{
euiss_cartogram_compare(
  sf_data,
  weight,
  types = c("cont", "dorling"),
  k_values = c(0.05, 0.1, 0.2),
  save_results = FALSE,
  output_dir = "cartograms",
  verbose = TRUE
)
}
\arguments{
\item{sf_data}{A Simple Features (sf) object.}

\item{weight}{The weight variable name (unquoted).}

\item{types}{Vector of cartogram types to create.}

\item{k_values}{Vector of k values to test (for applicable methods).}

\item{save_results}{Logical. Should results be saved to files?}

\item{output_dir}{Directory to save results if save_results=TRUE.}

\item{verbose}{Logical. Print progress information?}
}
\value{
List of cartogram sf objects with descriptive names.
}
\description{
This function creates several cartogram variants using different methods
and parameters, useful for exploring which approach works best for your data.
}
\examples{
\dontrun{
countries <- euiss_gisco(res = "20") \%>\% slice_head(n = 5)
countries$pop <- runif(nrow(countries), 1e6, 1e8)

cartograms <- euiss_cartogram_compare(
  countries, 
  weight = pop,
  types = c("cont", "dorling")
)

# Access individual cartograms
continuous_cartogram <- cartograms$cont_k005
dorling_cartogram <- cartograms$dorling_k005
}

}
