% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{euiss_create_basemap_plot}
\alias{euiss_create_basemap_plot}
\title{Create the final basemap plot}
\usage{
euiss_create_basemap_plot(
  countries,
  raster,
  roads,
  urban,
  urban_labels,
  lakes,
  rivers,
  admin,
  target_country,
  ras_name,
  crs,
  colors,
  show_graticule
)
}
\arguments{
\item{countries}{sf object with country boundaries}

\item{raster}{Raster data as tibble (optional)}

\item{roads}{sf object with roads (optional)}

\item{urban}{sf object with urban areas (optional)}

\item{urban_labels}{sf object with labels (optional)}

\item{lakes}{sf object with lakes (optional)}

\item{rivers}{sf object with rivers (optional)}

\item{admin}{sf object with admin boundaries (optional)}

\item{target_country}{Name of target country}

\item{ras_name}{Name of raster layer}

\item{crs}{Coordinate reference system}

\item{colors}{Named list of colors}

\item{show_graticule}{Show lat/lon grid?}
}
\description{
Create the final basemap plot
}
