% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-data.R
\name{euiss_download_gisco}
\alias{euiss_download_gisco}
\title{Download GISCO data from Eurostat}
\usage{
euiss_download_gisco(
  res = "60",
  year = "2024",
  type = "countries",
  dest_path = NULL,
  force = FALSE,
  timeout = 300
)
}
\arguments{
\item{res}{Resolution to download}

\item{year}{Year to download}

\item{type}{Data type to download}

\item{dest_path}{Destination path for downloads}

\item{force}{Re-download even if file exists}

\item{timeout}{Download timeout in seconds}
}
\description{
This function downloads GISCO geographic data directly from Eurostat's servers.
Note: This is a placeholder for future implementation of automated downloads.
}
\details{
Currently, this function provides download URLs and instructions for manual download.
Future versions will implement automated downloading.
}
\examples{
\dontrun{
# Get download information
euiss_download_gisco(res = "20", year = "2024")
}

}
