% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-data.R
\name{euiss_gisco_kosovo}
\alias{euiss_gisco_kosovo}
\title{Create Kosovo-compatible GISCO data}
\usage{
euiss_gisco_kosovo(
  res = "10",
  year = "2024",
  data_path = NULL,
  save_file = NULL
)
}
\arguments{
\item{res}{GISCO resolution to use}

\item{year}{GISCO year to use}

\item{data_path}{GISCO data path}

\item{save_file}{Optional filename to save the result}
}
\value{
sf object with Kosovo-compatible countries data
}
\description{
This function creates a version of GISCO countries data that includes Kosovo
as a separate entity, extracted from Serbia's boundaries using Natural Earth data.
}
\examples{
\dontrun{
# Create Kosovo-compatible data
countries_kosovo <- euiss_gisco_kosovo(res = "10")

# Save for future use
countries_kosovo <- euiss_gisco_kosovo(
  res = "10", 
  save_file = "countries_2024_10m_kosovo.rds"
)
}

}
