% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{euiss_reverse_geocode}
\alias{euiss_reverse_geocode}
\title{Reverse geocoding (REMOVED - Not Implemented)}
\usage{
euiss_reverse_geocode(...)
}
\arguments{
\item{...}{All parameters (ignored)}
}
\value{
Error message
}
\description{
This function has been removed as it was not fully implemented.
}
\details{
For reverse geocoding functionality, consider using:
\itemize{
  \item \code{tmaptools::rev_geocode_OSM()}
  \item \code{tidygeocoder::reverse_geocode()}
}
}
\keyword{internal}
